Update: Separated different routes into its own files, Update: DB constant into its own file
