+++ app/__init__.py
@@ -1,5 +1,4 @@
 from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
 from flask_jwt_extended import JWTManager
 from config import Config
 from .sample import sample_bp
@@ -7,8 +6,10 @@ from .donors import donors_bp
 from .donations import donations_bp
 from .campaigns import campaigns_bp
 from .campaign_donors import campaign_donors_bp
+from .auth import auth_bp
+from .db import db
 
-db = SQLAlchemy()
+
 jwt = JWTManager()
 
 
@@ -19,21 +20,16 @@ def create_app(config_class=Config):
     db.init_app(app)
     jwt.init_app(app)
 
-    from app.routes import auth_bp
 
     @app.route("/")
     def hello_world():
         return "Hello, World!"
 
-    app.register_blueprint(auth_bp, url_prefix="/api")
+    app.register_blueprint(auth_bp, url_prefix="/auth")
     app.register_blueprint(sample_bp, url_prefix="/sample")
     app.register_blueprint(donors_bp, url_prefix="/donors")
     app.register_blueprint(donations_bp, url_prefix="/donations")
     app.register_blueprint(campaigns_bp, url_prefix="/campaigns")
     app.register_blueprint(campaign_donors_bp, url_prefix="/campaign_donors")
-
-
-
-
 
     return app
