+++ app/auth/__init__.py
@@ -1,0 +1,45 @@
+from flask import Blueprint, jsonify, request
+from flask_jwt_extended import (
+    create_access_token,
+    create_refresh_token,
+    jwt_required,
+    get_jwt_identity,
+)
+from app.models import User
+
+auth_bp = Blueprint("auth", __name__)
+
+@auth_bp.route("/login", methods=["POST"])
+def login():
+    data = request.get_json()
+
+    if not data or "email" not in data or "password" not in data:
+        return jsonify({"error": "Missing email or password"}), 400
+
+    user = User.query.filter_by(email=data["email"]).first()
+
+    if user and user.check_password(data["password"]):
+        access_token = create_access_token(identity=user.id)
+        refresh_token = create_refresh_token(identity=user.id)
+        return (
+            jsonify({"access_token": access_token, "refresh_token": refresh_token}),
+            200,
+        )
+    return jsonify({"error": "Invalid credentials"}), 401
+
+
+@auth_bp.route("/protected", methods=["GET"])
+@jwt_required()
+def protected():
+    current_user_id = get_jwt_identity()
+    return jsonify({"message": f"Protected data for user {current_user_id}"}), 200
+
+
+@auth_bp.route("/refresh", methods=["POST"])
+@jwt_required(refresh=True)
+def refresh():
+    current_user_id = get_jwt_identity()
+    access_token = create_access_token(identity=current_user_id)
+    return jsonify({"access_token": access_token}), 200
+
+
