+++ app/models.py
@@ -1,4 +1,4 @@
-from app import db
+from .db import db
 from werkzeug.security import generate_password_hash, check_password_hash
 
 class User(db.Model):
