+++ t.txt
@@ -1,0 +1,8 @@
+db in __init__
+db (auth) imported in User
+user (auth) imported in Auth Route
+routes imported in __init__
+
+== Circular import detected ==
+
+
